gstreamer: add linbice dependency
This enables the nice plugins and also adds the webrtcbin plugins from
gst-plugins-bad.

Note, that this is different from the disabled libnice subproject. The
subproject clones libnice and try to install it in the system, which we don't
want. In contrast, adding the dependency just allows plugins to depend on the
system installed one.